Feature/product

+++ database/migrations/2024_04_05_100636_create_taxes_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('taxes', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->decimal('rate', 5, 2);
+            $table->integer('rate');
             $table->enum('status', ['active', 'inactive'])->default('active');
             $table->timestamps();
         });
